sys-apps/nvme-cli: Install udev rules and systemd units in correct places
Udev rules should be installed in rules.d directory (so in
`/usr/lib/udev/rules.d`), currently they are installed in
`/usr/lib/udev`. This mistake probably comes from a build system
change, where for old build system (in version 1.16), UDEVDIR was
supposed to be specified, and the build system defined UDEVRULESDIR
based on the former (see
https://github.com/linux-nvme/nvme-cli/blob/v1.16/Makefile#L17-L18). The
new build system expects the udev rules directory to be passed as a
parameter.

Similar mistake happened for the systemd unit directory - the systemd
unit files are installed in `/usr/lib/systemd/system`, not in
`/usr/lib/systemd`. This one stems from a change in meaning of the
`systemddir` option in the build system - in the old build system it
meant to be `/usr/lib/systemd` and the `system` subdirectory was
appended by the old build system when installing the units (see
https://github.com/linux-nvme/nvme-cli/blob/v1.16/Makefile#L142-L144).
The new build system does not do it any more - it just installs the
units to `systemddir` (see
https://github.com/linux-nvme/nvme-cli/blob/v1.16/Makefile#L142-L144),
thus making the `systemddir` option name rather misleading - probably
should be named `systemdunitdir`, as `systemddir` is not used for
anything else (see
https://github.com/linux-nvme/nvme-cli/blob/v2.4/Makefile#L49).

Signed-off-by: Krzesimir Nowak <knowak@microsoft.com>
Closes: https://github.com/gentoo/gentoo/pull/31159
Signed-off-by: Mike Gilbert <floppym@gentoo.org>