media-gfx/ueberzugpp: add 2.8.5
Adds USE=sway, may seem like it should be "wayland" with these
dependencies but afaik wayland has no real support for this and
at best it will be specific to the window manager.

Not adding a RDEPEND on sway given it may work with alternate
sway-based WMs and is more of an optfeature if support is built.

Don't use sway so I have not actually tried it (primarily use
kitty protocol for this which works anywhere kitty runs).

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>