autotools.eclass: Downgrade eqawarn for renaming configure.in
At this point, almost all upstreams will have switched to configure.ac.
Therefore, configure.in is most likely an indication of an inactive
upstream, and there is no reasonable way for the ebuild maintainer to
silence the warning (other than the ebuild renaming the file).

Keep the message as einfo, so there is still an indication that the file
was renamed.

Bug: https://bugs.gentoo.org/426262
Reviewed-by: Sam James <sam@gentoo.org>
Reviewed-by: Pacho Ramos <pacho@gentoo.org>
Signed-off-by: Ulrich Müller <ulm@gentoo.org>