x11-drivers/nvidia-drivers: migrate latest to linux-mod-r1
If no notable issues, the others will be done straight-to-stable
at a later date to avoid extra churn managing old branches.

This also renames USE=driver to USE=modules to be in-line with
other ebuilds, may disrupt a few users in chroots and such
(assuming didn't already have a global -modules) but hopefully is
obvious enough.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>