x11-drivers/nvidia-drivers: drop modules dir rename comment
Don't think will touch this for the time being given it
needs a postinst to delete old modules if they changed
location.

Not something was really thinking about, not a major issue
with most modules given will be cleaned on kernel upgrades
but with nvidia it means API mismatch issues until then.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>