media-libs/gst-plugins-good: fix build on systems with qt6
On systems with Qt6 installed gstreamer will automatically build the qt6
plugin. Qt in Gentoo however is not multilib enabled, which causes the
build to fail. Disable the module for now and once Qt6 is stable a
separate package for qt6 can be made (similar to the qt5 one in guru).

Signed-off-by: Nicolas Werner <nicolas.werner@hotmail.de>