media-libs/gst-plugins-bad: avoid automagic dep on libva
As noted on the bug, I need to discuss properly with leio whether or not the
split for -va makes sense, but -va is an orphan/leaf package right now so it's
lower priority than just fixing -bad which many people have installed. So, for now,
just backport the relevant upstream patches to fix the automagic for this side.

Bug: https://bugs.gentoo.org/907481
Bug: https://bugs.gentoo.org/907479
Signed-off-by: Sam James <sam@gentoo.org>