sys-apps/busybox: ignore cp -n error in postinstall
Before coreutils-9.2, cp -n (noclobber) would return success even if
skipping files.  Now it returns an error.  See also
https://debbugs.gnu.org/cgi/bugreport.cgi?bug=62572

This is used in postinstall with the purpose of not replacing already
existing utilities with links to busybox, so failing here is expected.
Unfortunately we cannot distinguish this from other errors.

Closes: https://bugs.gentoo.org/907432
Closes: https://github.com/gentoo/gentoo/pull/31257
Signed-off-by: Viorel Munteanu <ceamac@gentoo.org>