app-emulation/vkd3d-proton: USE=-debug -> global USE=strip in live
Will update the old ebuilds eventually but given this triggers
a rebuild with --changed-use (default enabled), will wait till
a bump and maybe stable to give a chance for people to update.

Note that USE=debug is still used to enable tracing here, part
of why these should ideally be separate flags.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>