x11-drivers/nvidia-drivers: use sandbox.d for /dev/nvidiactl+/dev/char
/dev/nvidiactl been a long standing issue, sometime appearing in sneaky
ways when a revdeps is built with opencl/cuda support even though the
package itself does not use it.

And /dev/char is newly needed with >=nvidia-drivers-525.105.17 or
>=535.43.02, but not 530.41.03. The production branch's 525.105.17
is newer than ~arch's long-living 530 and led to this being overlooked
until it hit stable (older stable 525.89.02 was not affected) and
was unaware of this until rebuilt libomp[offload] with 535 today
(note that 535.43.02 is unkeyworded, it's a beta).

Need /dev/char rather than /dev/char/195:255 given it tries to remove
+ create a symlink and does not simply try to write there.

This is not meant to be a full coverage of nvidia devices and only
for those being a widespread problem. Special needs or addwrite
(typically to run tests) should be handled manually or using
cuda.eclass' cuda_add_sandbox.

Adding /dev/char to all versions even if not needed *yet* just so it's
not overlooked when nvidia spreads it to other branches (except 390
given it's EOL, not to mention has no cuda packages anymore).

Bug: https://bugs.gentoo.org/904292
Bug: https://bugs.gentoo.org/905436
Closes: https://bugs.gentoo.org/904944
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>