media-libs/libglvnd-1.6.0: port to Darwin.
This commit ports media-libs/libglvnd-1.6.0 to Darwin. This
is implemented by removing -Wl,-Bsymbolic and -Wl,--version-script
from the linker option in a patch, and passing the build-time
option -Dasm=disabled to disable incompatible assembly code.

Porting libglvnd helps to advance the eventual goal of possibly
porting Mesa to Darwin and at least making software rendering
work, which is useful as Mesa is a Dependency of many packages.

Signed-off-by: Yifeng Li <tomli@tomli.me>