x11-drivers/nvidia-drivers: adjust openssl handling in 0/535
1. Noticed portage does not easily figure out rebuilds if do a plain
`emerge openssl:0/1.1`, but does if drop the || ( ) block. Not fully
correct to given wouldn't work with theoretical new subslots, but
dropping it for user experience sake (ultimately this is temporary).

Skip rebuild given this version won't be keyworded (this is for the
next one) and is still rebuilt on kernel bumps either way.

2. Had modified the .manifest block to handle this, but there is
no need to. Revert and use skip_files.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>