vala.eclass: workaround Vala breakage with Clang 16
Conditionally pass -Wno-incompatible-function-pointer-types for now in vala_setup
until Vala itself is fixed.

While this will affect other stuff in the same package, it's limited to when
Vala is enabled, so it's better than before at least. Right now, the situation
is both difficult for users & also blocks further testing.

Thanks to cyrillic for the nudge.

Bug: https://bugs.gentoo.org/892708
Signed-off-by: Sam James <sam@gentoo.org>