sys-apps/portage: unconditionally depend on sec-keys/openpgp-keys-gentoo-release
It's useful for people to have it installed and this keeps emerge-webrsync
working even with USE=-rsync-verify. The keys are tiny and have no dependencies
themselves, so I don't see the value in trying to mangle the script with sed
to default to --no-pgp-verify or similar.

(It'd be different if we had a proper build system which would let us do it. Maybe.)

Closes: https://bugs.gentoo.org/905868
Signed-off-by: Sam James <sam@gentoo.org>