net-im/telegram-desktop: Bump to 4.8.3
A very """minor""" upgrade, user-interface wise, but an annoying one
from the packaging side of things.

A new dependency to cppgir appeared, with no regards to packaging. This
applies to cppgir itself as well, as it bundles another library.

Dropped keywords, will need to be re-instated as the new dependencies
gain support for additional architectures.

D-Bus support was dropped:
https://github.com/telegramdesktop/tdesktop/commit/0a011db483eaaf4f9eec33c6af77d3bdf049c816
(I don't wish to go against this flow and re-introduce glib support
without dbus)

Signed-off-by: Esteve Varela Colominas <esteve.varela@gmail.com>
Closes: https://github.com/gentoo/gentoo/pull/31373
Signed-off-by: Sam James <sam@gentoo.org>