dev-libs/ntl: another AVX FFT build fix, this time on amd64.
Even on amd64, detection of the CPU features needed for AVX FFT can
fail. In particular, it can fail when -march=native is absent from
CXXFLAGS. Since the upstream documentation calls the AVX FFT
experimental, we disable it entirely. We also bring back NATIVE=on,
which should somewhat hide the CXXFLAGS issue until it can be addressed
upstream.

Closes: https://bugs.gentoo.org/815775
Signed-off-by: Michael Orlitzky <mjo@gentoo.org>