app-emulation/virtualbox: add 9999
Add live version.  It compiles the latest upstream public svn.
This ebuild also compiles the kernel modules, does not need on vbox-modules.
Upstream added support for python 3.12; I haven't checked it yet.
USE=doc does not work for now.
This version is not meant to be used, might be very unstable.

Signed-off-by: Viorel Munteanu <ceamac@gentoo.org>