python-utils-r1.eclass: Use @a instead of declare check for array
Use bash 5's @a variable substitution instead of `declare -p` check
to determine if `PYTHON_COMPAT` is an array.  This avoids a subshell,
and effectively makes the function roughly 3.5x faster:

```
 * Timing _python_set_impls
real 4854 it/s
user 4862 it/s
```

In practice, this roughly reduces total sourcing time for `dev-python/*`
from 57-58 s to 54 s.

Signed-off-by: Michał Górny <mgorny@gentoo.org>
Closes: https://github.com/gentoo/gentoo/pull/31454
Signed-off-by: Michał Górny <mgorny@gentoo.org>