media-sound/helm: Fix build on musl
Fixes three different issues with the bundled copy of JUCE in helm.
First, it conditions the usage of execinfo.h behind __GLIBC__ being
defined when on linux. Second, it removes the usage of lfs64 interfaces
and defines _FILE_OFFSET_BITS=64. Third, it conditions the usage of
_NL_IDENTIFICATION_LANGUAGE and _NL_IDENTIFICATION_TERRITORY behind
ifdefs, since musl does not define them.

Signed-off-by: Violet Purcell <vimproved@inventati.org>
Closes: https://github.com/gentoo/gentoo/pull/31463
Signed-off-by: Sam James <sam@gentoo.org>