dev-python/PyQt6: remove gcc dependency, update hack
Turns out that this *can* use clang++ but only if qtbase was built with
clang++ as well, otherwise it still uses g++ even with CXX=clang++

But given the former, there is little reason to enforce the dependency.

Ideally would use CXX but that would require more elaborate testing to
know what it is and what qtbase was built with (perhaps if ever add the
hack to qmake-utils.eclass).

No revbump given PyQt* bindings are a slowish build and depcleaning
gcc is a bit niche, the few interested can handle this manually until
next bump.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>