sci-libs/linux-gpib-modules: migrate to linux-mod-r1.eclass
Upstream has one target for all modules, this doesn't really
work with linux-mod(-r1) src_compile so we keep the workaround
that was already in the ebuild.

Now using MODULES_KERNEL_MIN to check the kernel version, got
rid of the custom solution in the ebuild.

Call modules_post_process to strip, compress and sign the modules

Signed-off-by: Andrew Ammerlaan <andrewammerlaan@gentoo.org>