games-emulation/pcsx2_patches: new package, add 0_p20230623 + 9999
This is distributed with PCSX2's appimage but tracked in a separate
repo. Simpler to handle as a separate package as bump may not be needed
at same time and updates tend to be non-urgent + game-specific (and
all patches are disabled by default).

PCSX2 newly gives a warning if patches.zip is missing (formerly
cheats_ws.zip, but now using its own patches interface), and will look
for it on the system's resources path. So this gives more reasons to
package this.

Provide 9999 as well given this does not get releases and users may
want fixes earlier than we do snapshots (albeit ~/.config/PCSX2/patches
or so should work afaik).

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>