sys-devel/gettext-0.22-r1: take more future-proof route against libintl
Patching configure only is brittle (what if one autoreconfs) and the
whole of it is wired to insist on building libintl (upstreams says in
order to test) so, instead of trying to avoid building libintl (yes,
it's a shame we have to waste cycles on it) just remove the library
after install.  This should be cleaner given we have no patch to
maintain (that can get outdated) and just makes the implementation a
one-liner.

Signed-off-by: Fabian Groffen <grobian@gentoo.org>