app-emulation/dxvk: update -mno-avx comment to match wine
Unsure for the status of issues with dxvk, the old Tk-Glitch issue
(who were building this with -march=native) is 404 given repos
changed around and unsure what the reproducer is. But given seemingly
bad code is being generated for recent Wine I think this is unsafe
enough that should keep it for now.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>