x11-drivers/nvidia-drivers: update and install vulkan icd for USE=-X
Not really tested, but it should be harmless vs not having the file.

Skipping revbump, letting it propagate on kernel upgrades / bumps
(470 and 525 are bumped in next commits). Please rebuild if want this
now for other versions.

For 390.x the icd is a template, while the newer drivers had hardcoded
libGLX, but docs mentions libEGL is usable for all versions (unlikely
to be useful for 390.x given its partial wayland support is broken).

Closes: https://bugs.gentoo.org/909181
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>