elisp-common.eclass: New function elisp-make-site-file
This creates and installs a site-init file for the package. The default
case will add the package's SITELISP subdirectory to Emacs' load-path.
This allows to remove all trivial site-init files (consisting only of a
load-path line).

More complicated cases can be handled by providing additional arguments,
which are appended as lines to the site-init file.

Signed-off-by: Ulrich Müller <ulm@gentoo.org>