app-emulation/virtiofsd: Install 50-qemu-virtiofsd.json
With the old virtiofsd provided by app-emulation/qemu the
corresponding JSON file descriptor is installed by QEMU. But with
this new rust version, we have to install it ourselves in order
for libvirt to pick it up. To avoid conflicts with the file
installed by app-emulation/qemu we can install it under a
different name. And because smaller number means higher priority
let's go with "40-qemu-virtiofsd.json" so that users with
<app-emulation/qemu-8.0.0 can benefit from rust version of the
daemon (if both are installed).

Closes: https://bugs.gentoo.org/909372
Signed-off-by: Michal Privoznik <michal.privoznik@gmail.com>