media-libs/libplacebo: add 6.287.0_rc1 (unkeyworded)
mpv-9999 now depends on >=5.266 (with USE=libplacebo) which is
unreleased bare this release candidate. Add to offer the option
to skip tracking libplacebo-9999 as well and prepare version
bounds given this seems incompatible with old mpv.

Note may not punctually bump further _rc if they are released,
please file a bug if you want one.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>