dev-vcs/breezy: add 3.3.0, enable py3.11, restrict tests
I've given up on adding tests. Yes, we have 29K out of 30K passed.
The 1K failing is ERROR/FAIL, so some of them might be false
positives, maybe something with portage sandbox, maybe with not CI
environment - I've given up. I conclude that it is good enough for
now.

Closes: https://bugs.gentoo.org/896944
Closes: https://bugs.gentoo.org/903674
Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>