x11-drivers/nvidia-drivers: drop additional framebuffer/drm warnings
Warnings were mostly added to help the transition to newer
kernels for unsuspecting users. May have helped in some cases,
not in others.

But with 6.1.x being stable for a while, there's little reason
to keep this wall of warnings *here* and try to keep it accurate
and updated -- especially when we can't tell what's really in-use
or what the user needs (this was just vague suggestions).

For initial setting up issues, it sounds better to refer to the
Wiki. So if anyone has anything to share with their experience
with FB (or other issues) feel free to edit it and improve it so
it can help others.

Also drop the "builtin" nouveau check that was part of this block.
Module is already blacklisted and, if users went out of their way
to make it builtin, then let's assume they know what they're doing.

Closing #910058 but rather than a fix it's more of a dissociation.

Closes: https://bugs.gentoo.org/910058
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>