app-misc/gramps: switch from distutils-r1 to python-single-r1
Gramps uses a heavily customised distutils-based setup.py which was
already somewhat fragile while using the legacy distutils-r1 mode
and which simply does not work correctly in PEP-517 mode. Just let
python-single-r1 handle the plumbing and call setup.py directly
as needed. Not only does this mean one package less in the tree to
depend on the legacy distutils-r1 but in the end, not having to
make the custom setup.py play nicely with that eclass actually makes
the ebuilds a bit simpler.

Closes: https://bugs.gentoo.org/909880
Signed-off-by: Marek Szuba <marecki@gentoo.org>