profiles: unmask Rust 1.70
If you still have issues with Rust 1.70, please file a Gentoo bug with all of
the details.

I've written up what the situation is on the Rust bug [0], but pasting it inline:
"""
Okay, so I spoke to some people internally and it looks like it was a combination of:

    Typo in the initial don't-strip-this-single-file
    Someone reporting a problem with that
    Adding the total strip-disabling and people getting confused about when the issue happens? (possibly not having picked up the previous fix?)
    The fact that llvm-strip seemingly didn't/doesn't help making it sound like another problem existed.

So, my plan now is:

    wait for someone who previously had an issue on ppc64;
    just totally disable stripping for the time being so we can get this out;
    make sure any issues get reported to us and then upstream if appropriate (rather than comments on IRC etc where it's easy to get confused about what someone's env is);
    re-evaluate disabling stripping for certain files instead if nothing comes up from the previous point.
"""

That feels to me to be a reasonable/plausible timeline of events.

matoro's since done that PPC64 testing and not hit any problems; ionen's been
using 1.70 for a while; I've rebuilt all rust pkgs (and used some of them) on
rust{,-bin}-1.70 machines without incident.

Thank you to matoro and ionen for helping me muddle my way through here.

So, all that said, let's unmask it and handle any new issues (although I'm
not expecting any) as-and-when/if they come in.

[0] https://github.com/rust-lang/rust/issues/112286#issuecomment-1630085688

Bug: https://github.com/rust-lang/rust/issues/112286
Signed-off-by: Sam James <sam@gentoo.org>