app-emulation/libvirt: Provide missing patch
In 45065ac9faa6 the libvirt-9.5.0-fix-paths-for-apparmor.patch
was renamed to libvirt-9.4.0-... but the live ebuild still refers
to the 9.5.0. Ideally, the live ebuild would be just made to
apply the renamed patch, but there was some more movement in the
upstream since and thus the patch doesn't apply cleanly anyways.
Therefore, rebase it and name it libvirt-9.6.0-... since the
upstream development was made after 9.5.0 release.

Signed-off-by: Michal Privoznik <michal.privoznik@gmail.com>