dev-python/setuptools: Move dev-python/wheel to RDEPEND
dev-python/wheel is needed for setuptools to be able to construct
wheels.  While this isn't a critical feature, it is useful for all
practical purposes and having to add the build-time dependency
everywhere is cumbersome.  This is especially a problem for the few
packages that use a custom wrapper over setuptools, and therefore
can't use the eclass logic for that.

Signed-off-by: Michał Górny <mgorny@gentoo.org>