dev-cpp/simpleini: add 4.20
Upstream added a CMakeLists but it does not really do anything for us
(no install target, no tests handling, no .cmake to install), and
is intended for in-source usage. So no changes in that regard.

However the normal Makefiles did get some improvements, there's
an install target and tests' now use pkg-config to find gtest.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>