media-libs/libva: Fix build with LLD 17
Upstream PR: https://github.com/intel/libva/pull/732

In the upcoming LLD 17 (and previously in LLD 16.0.0, but reverted in 16.0.1),
--no-undefined-version is becoming the default behavior. This causes the
configure check for --version-scripts support to fail due to the symbols in
the version script not being defined, which will cause the version script to
not be used, and the build will fail due to --no-undefined-version. This patch
adds '-Wl,--undefined-version' to the args of the configure check.

Signed-off-by: Violet Purcell <vimproved@inventati.org>