www-client/qutebrowser: update live
Now defaults to Qt6 and auto-fallbacks to Qt5 if missing or broken,
but there is little reason to use automagic in our case and it should
use the backend ebuild set dependencies for (users can still override
this at runtime if really wanted).

Using sed is what machinery.py suggests.

On a side-note, selected qt6 tests currently pass and qt5 has one
failure. Will be reviewed further on release.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>