media-video/mpv: backport fix for use with pipewire-0.3.75
Albeit, regardless of this patch or stable pipewire, ao=pipewire on
0.35.1 seems to have several other issues (in my VM it results in
horribly crackling audio), and I personally recommend ao=pulse
(or jack/alsa) instead to use with pipewire. Pipewire does work
great in mpv-9999, so next release should be better.

Considered a ~arch revbump, but given this only touch pipewire
code and that ao=pipewire is already a bit bad on 0.35.1 I'm not
sure it's worth the stabilization trouble (aka current keywords
are intentional).

Tested with both 0.3.75 and stable 0.3.71, but will keep -r1 for
a bit in case of something I missed.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>