media-video/mpv: add 0.36.0
Neither me nor pipewire upstream expected a release soon, but
turns out did not have to worry about that. Also includes more
pipewire fixes than just the emergency one so ao=pipewire users
should upgrade ahead of stable.

Note that this needs the latest version of libplacebo or else
gpu-next won't be enabled (that version is however incompatible
with <ffmpeg-6.1 and <vlc-4 -- not that it's currently enabled in
Gentoo for vlc-3 and is off-by-default on ffmpeg).

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>