dev-python/PyQtWebEngine: enable py3.12, use distutils-r1+PEP517
Also add the missing dependencies on qtnetwork:5 and qtwidgets:5
while revbumping. They technically cannot be missed as they are
pulled by PyQt5[network,widgets], but this is to be proper given
it's directly linked with these.

Note these could be optional (like they are in PyQt6-WebEngine behind
IUSE=+widgets), but not going to bother changing this at this point
unless someone needs this.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>