www-client/qutebrowser: add note wrt py3.12 in live
Last blocker is pyside2+6 to enable py3.12 on QtPy which is needed
by pytest-qt, but tests on pyside are restricted and I do not
use/follow either of these to be able to know if it works.

If ignore the pyside deps and force enable 3.12 just to be able
to run qutebrowser's tests, it passes fine.

Leaving this on hold for now so don't need to restrict tests.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>