media-video/ffmpeg: Rename AMF env.d file, and revbump
Leading digits are required for env-update to consider the file.

This commit leaves the file in the repository as is and uses "newenvd"
to install it under a different name. This makes sure that the 6.0-r2
ebuild continues to work. If that is not required, a simpler solution
would be to instead rename the file in the repository to
99amf-env-vulkan-override and keep using "doenvd".

Closes: https://bugs.gentoo.org/911249

Signed-off-by: Karl-Johan Karlsson <creideiki@lysator.liu.se>