dev-lang/php: improve testing in the 8.2 slot.
We had basically given up on the PHP test suite due to a high number of
"expected" failures. I've tried to track most of these down and
report/fix them upstream. In the meantime, all known failing tests have
been disabled one-at-a-time, with pointers to the upstream issue/PR. And
with that done, the test suite is once again run in anger; if it fails,
the installation will too.

Signed-off-by: Michael Orlitzky <mjo@gentoo.org>