dev-lang/go: fix bootstrap dependency & build-time detection
We need to both depend on a new enough bootstrap version for correct upgrade
ordering *and* check the version we use in src_compile rather than just picking
whichever one of go or go-bootstrap is installed.

So, add a minimum version variable and use it in BDEPEND & src_compile's has_version
calls.

Sync live while at it.

Closes: https://bugs.gentoo.org/911275
Signed-off-by: Sam James <sam@gentoo.org>