dev-lang/php: disable USE=coverage and hide /usr/bin/php.
Code coverage is "FOR DEVELOPERS ONLY!!", and requires GCC. It's not
really something we want to support for end users, and it's contributing
to bug 900210, so now it's gone. The rest of that bug is caused by PHP's
./configure script detecting an already-installed PHP (at /usr/bin/php)
and running it.

Obviously PHP isn't needed to build itself, so allowing it to detect and
(apparently) run an existing installation is only multiplying the insane
number of configurations we already support. To avoid that in the future
(and to fix bug 900210 right now), we hide the system "php" by
overriding its cache variable during ./configure.

Closes: https://bugs.gentoo.org/900210
Signed-off-by: Michael Orlitzky <mjo@gentoo.org>