media-libs/rubberband: Restrict GCC dependency to ppc
The gcc RDEPEND is presumably due to the ebuild linking against libatomic.
However, this is conditioned behind either ppc or libstdc++, which the latter
would imply that GCC is already installed. This needlessly pulls in GCC on musl/llvm systems,
and making the gcc dependency ppc-only shouldn't cause any breakage.

Signed-off-by: Violet Purcell <vimproved@inventati.org>
Closes: https://github.com/gentoo/gentoo/pull/32073
Signed-off-by: Sam James <sam@gentoo.org>