toolchain-funcs.eclass: cleanup obsolete bits in tc-ld-force-bfd
I didn't want to drop this entirely in c26413ae8029f1557aae95b6f8744fc16671fa3b
as I wanted to keep it to just fixing the error I'd made, but I don't see the
point in keeping this now anyway - non-gcc/clang isn't a thing and if it is
I'd hope they'd support -fuse-ld=bfd anyway.

Plus it's an obvious approach if we need it for something else like forcing
an alt linker that isn't bfd anyway, no need for having cruft remaining.

Signed-off-by: Sam James <sam@gentoo.org>