sci-mathematics/pari: new revision to fix USE=X handling
Pari's build system is interesting, and choosing a graphics library (for
use in plotting) is... not straightforward. We have two patches that
affect it, and still nobody has noticed in all this time that USE=X does
precisely nothing at all.

I've just sent one of those patches, for building against fltk,
upstream. To avoid having to reroll the other patch and add an X11 case
to it, this revision switches from using e.g. the "--with-fltk" flag to
using "--graphic=fltk" instead. This turns out to be a stronger hint to
the build system, and more imporantly supports both none/X11 as options,
unlike --with-foo. Hopefully it allows us to go patch-free with the next
release of pari.

Signed-off-by: Michael Orlitzky <mjo@gentoo.org>