x11-drivers/nvidia-drivers: adjust IBT warning again
Hoping to be able to remove this from 535 branch at least, but
I have no way to really confirm so leaving it there for now.

Last time I removed this warning I just got someone on IRC with
issues who wondered why the ebuild didn't warn about this but
that was with an older version.

This aside, do disable warning with USE=kernel-open, pretty sure
that should always be fine.

Having a bug open to track this should be useful anyhow.

Bug: https://bugs.gentoo.org/911142
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>