dev-lang/go: add 1.20.7
We need to both depend on a new enough bootstrap version for correct upgrade
ordering *and* check the version we use in src_compile rather than just picking
whichever one of go or go-bootstrap is installed.

So, add a minimum version variable and use it in BDEPEND & src_compile's has_version
calls.

Thanks to sam@gentoo.org for the go bootstrap updates.

Closes: https://bugs.gentoo.org/911275
Closes: https://github.com/gentoo/gentoo/pull/32061
Signed-off-by: William Hubbs <williamh@gentoo.org>