app-emulation/wine-staging: tentative support for 32bit w/o multilib
This been possible to "build" since early wine-8 but was too
incomplete to be really useful and became more usable since roughly
8.10.

Scarecely tested, but was still able to get hardware accelerated
32bit wglgears.exe (like glxgears) on a no-multilib prefix install.
Not tested on musl yet, but alpine uses it and should work in theory.

Only going enable in current latest 8.13 and 9999 so don't expect
this on wine-proton/vanilla-8.0.x where it'd be near useless
(for wine-proton it likely will not be until wine-proton-9).

Does not feel at a stage where it should be default except on
no-multilib profiles (abi_x86_32 is still recommended and wine itself
outputs a warning about being experimental when using it), but in the
event primarily use 64bit while still needing some 32bit support for
e.g. bad installers, then this should be good enough.

Still requires USE=abi_x86_32 on mingw64-toolchain and wine-gecko
(also dxvk/vkd3d-proton if used), but these have no requirements
on multilib libraries of their own. In the future may potentially
use a different USE on these to avoid confusion and integrate better
with no-multilib profiles.

For convenience moved the CROSS{CC,*FLAGS} out of the loop -- meaning
this no longer tests flags separately for 64+32 but should not be an
issue.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>