app-emulation/wine-vanilla: fix build with clang:17
Was silently ignored with <clang-16, but clang:17 now considers this
an error.

Working -mabi=ms is required with USE=-mingw, but with USE=mingw seems
it gets used in install phase possibly(?) by mistake. As a quick fix,
drop the option for now. Prefer to leave alone for gcc, so done in
ebuild w/ tc-is-clang.

Also add an early abort for USE=-mingw while here, this was always
failing due to missing -mabi=ms even with older clang.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>