app-emulation/wine-proton: fix build with clang:17
Was silently ignored with <clang-16, but clang:17 now considers this
an error.

Working -mabi=ms would be required if --without-mingw, but with it
seems it gets used in install phase possibly(?) by mistake. As a
quick fix, drop the option for now. Prefer to leave alone for gcc,
so done in ebuild w/ tc-is-clang.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>