eclass/ruby-fakegem.eclass: always avoid minitest plugins
Minitest will by default scan for any plugin installed by any gem and
register it. This has been a source of test failures when a plugin can
no longer be initialized for some reason.

Many ebuilds already set the environment variable for this, but we never
want autoloading so turn this off in the eclass instead.

Signed-off-by: Hans de Graaff <graaff@gentoo.org>