app-emulation/wine-vanilla: pass -mno-avx with mingw unconditionally
Upon further consideration 9bef96cec566f1e06f5e9c40e52785b7b9702afa
result in rather unexpected behavior even if we consider that
USE=custom-cflags is unsupported, and giving a way to skip -mno-avx
may not be all that worth it.

So revert plus tidy and add this bugref.

Closes: https://bugs.gentoo.org/912268
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>