dev-cpp/abseil-cpp: force use of C++17 delegation in absl classes
Starting with abseil-2023xxyy abseil will start to delegate various absl:: types
to their std:: equivalents by default when built with >=std=c++17.
This older release still defaults to creating an absl/base/options.h file where the
delegation is configured to happen at build time, creating ABI problems for consumers.
In order to be consistent with future releases we therefore enforce the delegation to
std::* implementations in this release as well.
This implies that any consumers of abseil - directly or indirectly - will need
to be built with >= -std=c++17.

Closes: https://bugs.gentoo.org/912280
Signed-off-by: Holger Hoffstätte <holger@applied-asynchrony.com>