dev-db/redis: add 7.2.0
Two patches were necessary to rebase, the *-shared.patch was renamed to
*-system-jemalloc.patch which better suits the change.

Boost-1.0 was appended to LICENSE because of newly bundled fpconv.

It is not necessary to append -DENABLE_CJSON_GLOBAL to cflags as it is
correctly set in deps/Makefile, thus flag-o-matic is no longer used.

-O2 is stripped from LUA_CLFAGS in order to respect user's CFLAGS.

The "big" sed for all Makefiles was revised, all expressions are moved
to a local mysedconf array, obsolete expressions are removed and it
newly strips recently appended -Werror* flags.

myconf is transformed to an array, with newly added OPTIMIZATION and
DEBUG_FLAGS defined as empty strings in order to respect user's CFLAGS.
USE_SYSTEMD is moved to the array as well.

github link is appended to HOMEPAGE.

Signed-off-by: Petr Vaněk <arkamar@atlas.cz>
Closes: https://github.com/gentoo/gentoo/pull/32338
Signed-off-by: Sam James <sam@gentoo.org>