app-text/atril: add dev-libs/mathjax upper bound
Upstream requires the old mathjax-2.x (it loads MathJax.js), so when we
add mathjax-3.x, bad things will happen otherwise.

It's not clear to me that this ever worked, because the path to
MathJax.js is hard-coded upstream as /usr/share/javascript/mathjax,
and that's not what Gentoo uses -- but let's be safe.

Bug: https://bugs.gentoo.org/837722
Signed-off-by: Michael Orlitzky <mjo@gentoo.org>