www-client/qutebrowser: update/tidy live
No notable changes, but clarify/reword a few comments, call
python_gen_cond_dep less times, skip a failing test with Qt5,
and uncomment the Qt6 warning. Rest is just style.

This is in preparation for the imminent qutebrowser-3 release,
tests pass for both Qt5 and Qt6 in current state.

May still be other small adjustments to do depending on how
the release tarball is created.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>